Make Connections per-network.
Connection#networks (which we were not filling properly) is now Connection#network, with only a single network.
Update User#follow, User#unfollow to take a 2nd argument to specify the network(s) to follow the other user on.
Add User#followees, to get list of users that the user follows.
Add User#networks, to get list of networks that the user has accounts on.